+++ src/index.html
@@ -131,6 +131,9 @@
 
 <script src="js/resizer.js" charset="utf-8"></script>
 <script src="js/upload.js" charset="utf-8"></script>
-<script src="pictures.js"></script>
+<script src="js/main.js" charset="utf-8"></script>
+<!--<script src="js/load.js" charset="utf-8"></script>-->
+<!--<script src="js/get-picture-element.js" charset="utf-8"></script>-->
+<!--<script src="pictures.js"></script>-->
 </body>
 </html>
