+++ src/js/load.js
@@ -1,0 +1,16 @@
+module.exports = function () {
+  var load = function (url, callback, callbackName) {
+    if (!callbackName) {
+      callbackName = 'cb' + Date.now();
+    }
+
+    window[callbackName] = function (data) {
+      pictures = data;
+      callback(data);
+    };
+
+    var script = document.createElement('script');
+    script.src = url + '?callback=' + callbackName;
+    document.body.appendChild(script);
+  };
+};
