+++ src/js/main.js
@@ -1,1 +1,7 @@
 'use strict';
+
+var load = require('./load');
+var getPictureElement = require('./picture');
+var pictures = require('./pictures');
+
+console.log('hi');
